Program Restructure
Restructured the program to folow a nore nodular design, introducing main.py as a menu system used to access the photo frame. also added files for current intended features
