+++ modules/MusicApp.py
@@ -1,0 +1,24 @@
+import tkinter as tk
+from tkinter import ttk
+
+class MusicPlayer:
+    def __init__(self, parent):
+        self.parent = parent
+
+        # Initialize music and speaker module UI and logic
+
+    def toggle_music(self):
+        # Toggle music playback
+        pass
+
+    def adjust_volume(self, delta):
+        # Adjust speaker volume
+        pass
+
+    def close(self):
+        # Clean up and close module
+        pass
+
+    def show(self):
+        # Show module UI
+        pass
