+++ modules/VideoPlayer.py
@@ -1,0 +1,16 @@
+import tkinter as tk
+from tkinter import ttk
+
+class VideoPlayer:
+    def __init__(self, parent):
+        self.parent = parent
+
+        # Initialize video player UI and logic
+
+    def close(self):
+        # Clean up and close module
+        pass
+
+    def show(self):
+        # Show module UI
+        pass
